+++ cloud-siem-20220616/alibabacloud_cloud_siem20220616/models/_describe_dispose_and_playbook_request.py
@@ -7,38 +7,36 @@ from darabonba.model import DaraModel
 class DescribeDisposeAndPlaybookRequest(DaraModel):
     def __init__(
         self,
+        available_only: bool = None,
         current_page: int = None,
         entity_type: str = None,
         entity_uuid: str = None,
+        entity_uuid_list: str = None,
         incident_uuid: str = None,
         page_size: int = None,
         region_id: str = None,
         role_for: int = None,
         role_type: int = None,
     ):
-        # The page number. The value must be greater than or equal to 1.
+        self.available_only = available_only
+        # The current page number. The value must be greater than or equal to 1.
         self.current_page = current_page
         # The entity type. Valid values:
-        # - ip: IP address
-        # - process: process
-        # - file: file
         self.entity_type = entity_type
         # The entity UUID.
         self.entity_uuid = entity_uuid
-        # The event UUID.
+        self.entity_uuid_list = entity_uuid_list
+        # The incident UUID.
         self.incident_uuid = incident_uuid
         # The number of entries per page. Maximum value: 100.
         self.page_size = page_size
-        # The region in which the data management center of the threat analysis feature resides. Select the data management center based on the region of your assets. Valid values:
-        # - cn-hangzhou: Your assets reside in the Chinese mainland or Hong Kong (China).
-        # - ap-southeast-1: Your assets reside in regions outside the Chinese mainland.
+        # The region where the threat analysis data management center is located. Specify the management center region based on the region of your assets. Valid values:
+        # - cn-hangzhou: Your assets are located in the Chinese mainland or Hong Kong (China).
+        # - ap-southeast-1: Your assets are located outside China.
         self.region_id = region_id
         # The user ID of the member to which the administrator switches the view.
         self.role_for = role_for
-        # The view type. Valid values:
-        # 
-        # - 0: the view of the current Alibaba Cloud account.
-        # - 1: the view of all accounts in the enterprise.
+        # The view type.
         self.role_type = role_type
 
     def validate(self):
@@ -49,6 +47,9 @@ class DescribeDisposeAndPlaybookRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.available_only is not None:
+            result['AvailableOnly'] = self.available_only
+
         if self.current_page is not None:
             result['CurrentPage'] = self.current_page
 
@@ -57,6 +58,9 @@ class DescribeDisposeAndPlaybookRequest(DaraModel):
 
         if self.entity_uuid is not None:
             result['EntityUuid'] = self.entity_uuid
+
+        if self.entity_uuid_list is not None:
+            result['EntityUuidList'] = self.entity_uuid_list
 
         if self.incident_uuid is not None:
             result['IncidentUuid'] = self.incident_uuid
@@ -77,6 +81,9 @@ class DescribeDisposeAndPlaybookRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AvailableOnly') is not None:
+            self.available_only = m.get('AvailableOnly')
+
         if m.get('CurrentPage') is not None:
             self.current_page = m.get('CurrentPage')
 
@@ -85,6 +92,9 @@ class DescribeDisposeAndPlaybookRequest(DaraModel):
 
         if m.get('EntityUuid') is not None:
             self.entity_uuid = m.get('EntityUuid')
+
+        if m.get('EntityUuidList') is not None:
+            self.entity_uuid_list = m.get('EntityUuidList')
 
         if m.get('IncidentUuid') is not None:
             self.incident_uuid = m.get('IncidentUuid')