+++ cloud-siem-20220616/alibabacloud_cloud_siem20220616/models/_describe_dispose_strategy_playbook_request.py
@@ -13,25 +13,22 @@ class DescribeDisposeStrategyPlaybookRequest(DaraModel):
         role_type: int = None,
         start_time: int = None,
     ):
-        # The end time of the query. This value is a UNIX timestamp in milliseconds.
+        # The end time of the query, in milliseconds.
         # 
         # This parameter is required.
         self.end_time = end_time
-        # The region of the Data Management center for Threat Analysis. Select a region based on the location of your assets. Valid values:
-        # 
-        # - cn-hangzhou: The assets are in the Chinese mainland or China (Hong Kong).
-        # 
-        # - ap-southeast-1: The assets are outside China.
+        # The region where the data management center of threat detection and response is located. Select the management center based on the region of your assets. Valid values:
+        # - cn-hangzhou: the asset belongs to the Chinese mainland or Hong Kong (China).
+        # - ap-southeast-1: the asset belongs to a region outside China.
         self.region_id = region_id
-        # The user ID of the member. An administrator can use this parameter to switch to the member\\"s view.
+        # The ID of the user to which the administrator switches the view.
         self.role_for = role_for
-        # The view type.
+        # The view type. Valid values:
         # 
-        # - 0: The view of the current Alibaba Cloud account.
-        # 
-        # - 1: The view of all accounts in the enterprise.
+        # - 0: the view of the current Alibaba Cloud account.
+        # - 1: the view of all accounts in the enterprise.
         self.role_type = role_type
-        # The start time of the query. This value is a UNIX timestamp in milliseconds.
+        # The start time of the query, in milliseconds.
         # 
         # This parameter is required.
         self.start_time = start_time